backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-limiter==3.12
PyJWT==2.10.1
supabase==2.13.0
python-dotenv==1.1.0
requests==2.32.3
gunicorn==23.0.0
google-genai==1.7.0
numpy==2.2.3
cryptography==44.0.0
